requirements.txt
ipykernel==6.29.0
mlflow==2.10.0
pillow==10.2.0
requests==2.31.0
streamlit==1.30.0
torch==2.1.2
torchvision==0.16.2
transformers==4.37.1
optuna==3.5.0
pycocotools==2.0.7
schedule==1.2.0
plotly==5.18.0
tqdm==4.66.1
pandas==2.1.4
numpy==1.26.3
